Update requirements.txt
Added missing requirements that would be later on imported such as PyQt5 and below it.

Also referenced my own fork of the OpenVoice git so it could be downloaded automatically. My own fork has a few minor changes in requirements and installation process that I needed to make for it to work with python 3.11 in my case. It should still, howver, work with other versions. 
One major issue was that there were conflicting dependencies to begin with. This is resolved in my fork that is linked.